pyproject.toml
[project]
name = "game-library-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "django-stubs>=6.0.1",
    "djangorestframework>=3.17.1",
    "dotenv>=0.9.9",
    "gunicorn>=25.3.0",
    "markdown>=3.10.2",
    "psycopg2>=2.9.11",
    "whitenoise>=6.12.0",
]

[tool.django-stubs]
django_settings_module = "game_library_manager.settings"
